feat: Use u8 instead of usize to represent register index
This change has a ~10% performance boost since a smaller instruction
object means we can fit more instructions in CPU cache in one load.

+++ src/instructions/i.rs
@@ -147,15 +147,15 @@ impl Execute for Rtype {
             RtypeInstruction::OR => common::or(machine, self.rd, self.rs1, self.rs2),
             RtypeInstruction::AND => common::and(machine, self.rd, self.rs1, self.rs2),
             RtypeInstruction::SLL => {
-                let shift_value = machine.registers()[self.rs2].clone()
+                let shift_value = machine.registers()[self.rs2 as usize].clone()
                     & Mac::REG::from_usize(Mac::REG::SHIFT_MASK);
-                let value = machine.registers()[self.rs1].clone() << shift_value;
+                let value = machine.registers()[self.rs1 as usize].clone() << shift_value;
                 update_register(machine, self.rd, value);
             }
             RtypeInstruction::SLLW => {
                 let shift_value =
-                    machine.registers()[self.rs2].clone() & Mac::REG::from_usize(0x1F);
-                let value = machine.registers()[self.rs1].clone() << shift_value;
+                    machine.registers()[self.rs2 as usize].clone() & Mac::REG::from_usize(0x1F);
+                let value = machine.registers()[self.rs1 as usize].clone() << shift_value;
                 update_register(
                     machine,
                     self.rd,
@@ -163,15 +163,16 @@ impl Execute for Rtype {
                 );
             }
             RtypeInstruction::SRL => {
-                let shift_value = machine.registers()[self.rs2].clone()
+                let shift_value = machine.registers()[self.rs2 as usize].clone()
                     & Mac::REG::from_usize(Mac::REG::SHIFT_MASK);
-                let value = machine.registers()[self.rs1].clone() >> shift_value;
+                let value = machine.registers()[self.rs1 as usize].clone() >> shift_value;
                 update_register(machine, self.rd, value);
             }
             RtypeInstruction::SRLW => {
                 let shift_value =
-                    machine.registers()[self.rs2].clone() & Mac::REG::from_usize(0x1F);
-                let value = machine.registers()[self.rs1].zero_extend(&Mac::REG::from_usize(32))
+                    machine.registers()[self.rs2 as usize].clone() & Mac::REG::from_usize(0x1F);
+                let value = machine.registers()[self.rs1 as usize]
+                    .zero_extend(&Mac::REG::from_usize(32))
                     >> shift_value;
                 update_register(
                     machine,
@@ -180,15 +181,15 @@ impl Execute for Rtype {
                 );
             }
             RtypeInstruction::SRA => {
-                let shift_value = machine.registers()[self.rs2].clone()
+                let shift_value = machine.registers()[self.rs2 as usize].clone()
                     & Mac::REG::from_usize(Mac::REG::SHIFT_MASK);
-                let value = machine.registers()[self.rs1].signed_shr(&shift_value);
+                let value = machine.registers()[self.rs1 as usize].signed_shr(&shift_value);
                 update_register(machine, self.rd, value);
             }
             RtypeInstruction::SRAW => {
                 let shift_value =
-                    machine.registers()[self.rs2].clone() & Mac::REG::from_usize(0x1F);
-                let value = machine.registers()[self.rs1]
+                    machine.registers()[self.rs2 as usize].clone() & Mac::REG::from_usize(0x1F);
+                let value = machine.registers()[self.rs1 as usize]
                     .sign_extend(&Mac::REG::from_usize(32))
                     .signed_shr(&shift_value);
                 update_register(
@@ -198,14 +199,14 @@ impl Execute for Rtype {
                 );
             }
             RtypeInstruction::SLT => {
-                let rs1_value = &machine.registers()[self.rs1];
-                let rs2_value = &machine.registers()[self.rs2];
+                let rs1_value = &machine.registers()[self.rs1 as usize];
+                let rs2_value = &machine.registers()[self.rs2 as usize];
                 let value = rs1_value.lt_s(&rs2_value);
                 update_register(machine, self.rd, value);
             }
             RtypeInstruction::SLTU => {
-                let rs1_value = &machine.registers()[self.rs1];
-                let rs2_value = &machine.registers()[self.rs2];
+                let rs1_value = &machine.registers()[self.rs1 as usize];
+                let rs2_value = &machine.registers()[self.rs2 as usize];
                 let value = rs1_value.lt(&rs2_value);
                 update_register(machine, self.rd, value);
             }
@@ -230,21 +231,21 @@ impl Execute for Itype {
             ItypeInstruction::ORI => common::ori(machine, self.rd, self.rs1, self.imm),
             ItypeInstruction::ANDI => common::andi(machine, self.rd, self.rs1, self.imm),
             ItypeInstruction::SLTI => {
-                let rs1_value = &machine.registers()[self.rs1];
+                let rs1_value = &machine.registers()[self.rs1 as usize];
                 let imm_value = Mac::REG::from_i32(self.imm);
                 let value = rs1_value.lt_s(&imm_value);
                 update_register(machine, self.rd, value);
             }
             ItypeInstruction::SLTIU => {
-                let rs1_value = &machine.registers()[self.rs1];
+                let rs1_value = &machine.registers()[self.rs1 as usize];
                 let imm_value = Mac::REG::from_i32(self.imm);
                 let value = rs1_value.lt(&imm_value);
                 update_register(machine, self.rd, value);
             }
             ItypeInstruction::JALR => {
                 let link = machine.pc().overflowing_add(&Mac::REG::from_usize(4));
-                let mut next_pc =
-                    machine.registers()[self.rs1].overflowing_add(&Mac::REG::from_i32(self.imm));
+                let mut next_pc = machine.registers()[self.rs1 as usize]
+                    .overflowing_add(&Mac::REG::from_i32(self.imm));
                 next_pc = next_pc & (!Mac::REG::one());
                 update_register(machine, self.rd, link);
                 return Ok(Some(next_pc));
@@ -294,8 +295,8 @@ impl Execute for Stype {
 
 impl Execute for Btype {
     fn execute<Mac: Machine>(&self, machine: &mut Mac) -> Result<Option<Mac::REG>, Error> {
-        let rs1_value = &machine.registers()[self.rs1];
-        let rs2_value = &machine.registers()[self.rs2];
+        let rs1_value = &machine.registers()[self.rs1 as usize];
+        let rs2_value = &machine.registers()[self.rs2 as usize];
         let condition = match &self.inst {
             BtypeInstruction::BEQ => rs1_value.eq(&rs2_value),
             BtypeInstruction::BNE => rs1_value.ne(&rs2_value),
@@ -615,19 +616,19 @@ pub fn factory<R: Register>(instruction_bits: u32) -> Option<GenericInstruction>
                     0b_101 => Some(Instruction::CsrI(CsrIType {
                         csr,
                         rd,
-                        zimm: rs1_zimm as u32,
+                        zimm: u32::from(rs1_zimm),
                         inst: CsrIInstruction::CSRRWI,
                     })),
                     0b_110 => Some(Instruction::CsrI(CsrIType {
                         csr,
                         rd,
-                        zimm: rs1_zimm as u32,
+                        zimm: u32::from(rs1_zimm),
                         inst: CsrIInstruction::CSRRSI,
                     })),
                     0b_111 => Some(Instruction::CsrI(CsrIType {
                         csr,
                         rd,
-                        zimm: rs1_zimm as u32,
+                        zimm: u32::from(rs1_zimm),
                         inst: CsrIInstruction::CSRRCI,
                     })),
                     _ => None,
